backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
channels==4.3.1
channels-redis==4.3.0
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.1
django==5.2.4
django-channels==0.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.3.0
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
langextract==1.0.3
ml-collections==1.1.0
more-itertools==10.7.0
msgpack==1.1.1
multidict==6.6.3
numpy==2.3.2
oauthlib==3.3.1
openai==1.98.0
pandas==2.3.1
propcache==0.3.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==8.5.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
django-jazzmin==3.1.0
google-generativeai==0.8.3
anthropic==0.35.0
